move rest of the "administrator 2fa auth" functionality from project-base to the framework package
