+++ packages/framework/src/Model/Administrator/Mail/TwoFactorAuthenticationMail.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Model\Administrator\Mail;
+namespace Shopsys\FrameworkBundle\Model\Administrator\Mail;
 
-use App\Component\Setting\Setting;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
+use Shopsys\FrameworkBundle\Component\Setting\Setting;
 use Shopsys\FrameworkBundle\Model\Mail\MailTemplate;
 use Shopsys\FrameworkBundle\Model\Mail\MessageData;
 use Shopsys\FrameworkBundle\Model\Mail\MessageFactoryInterface;
@@ -13,20 +13,22 @@ use Shopsys\FrameworkBundle\Model\Mail\Setting\MailSetting;
 
 class TwoFactorAuthenticationMail implements MessageFactoryInterface
 {
-    public const TWO_FACTOR_AUTHENTICATION_CODE = 'two_factor_authentication_code';
-    public const VARIABLE_AUTHENTICATION_CODE = '{authentication_code}';
+    public const string TWO_FACTOR_AUTHENTICATION_CODE = 'two_factor_authentication_code';
+    public const string VARIABLE_AUTHENTICATION_CODE = '{authentication_code}';
 
     /**
-     * @param \App\Component\Setting\Setting $setting
+     * @param \Shopsys\FrameworkBundle\Component\Setting\Setting $setting
      * @param \Shopsys\FrameworkBundle\Component\Domain\Domain $domain
      */
-    public function __construct(private Setting $setting, private Domain $domain)
-    {
+    public function __construct(
+        protected readonly Setting $setting,
+        protected readonly Domain $domain,
+    ) {
     }
 
     /**
-     * @param \App\Model\Mail\MailTemplate $template
-     * @param \App\Model\Administrator\Administrator $administrator
+     * @param \Shopsys\FrameworkBundle\Model\Mail\MailTemplate $template
+     * @param \Shopsys\FrameworkBundle\Model\Administrator\Administrator $administrator
      * @return \Shopsys\FrameworkBundle\Model\Mail\MessageData
      */
     public function createMessage(MailTemplate $template, $administrator)
