+++ project-base/app/src/Model/Mail/MailTemplateConfiguration.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Model\Mail;
 
-use App\Model\Administrator\Mail\TwoFactorAuthenticationMail;
 use App\Model\Customer\Mail\CustomerActivationMail;
 use App\Model\Order\Mail\OrderMail;
 use Shopsys\FrameworkBundle\Model\Mail\MailTemplateConfiguration as BaseMailTemplateConfiguration;
@@ -22,7 +21,6 @@ class MailTemplateConfiguration extends BaseMailTemplateConfiguration
 
         $this->registerExtendedOrderStatusMailTemplates();
         $this->registerCustomerActivationMailTemplate();
-        $this->registerTwoFactorAuthenticationCodeMailTemplate();
     }
 
     /**
@@ -81,27 +79,5 @@ class MailTemplateConfiguration extends BaseMailTemplateConfiguration
     {
         return parent::createOrderStatusMailTemplateVariables()
             ->addVariable(OrderMail::VARIABLE_TRACKING_INSTRUCTIONS, t('Tracking instructions'), MailTemplateVariables::CONTEXT_BODY);
-    }
-
-    /**
-     * @return \Shopsys\FrameworkBundle\Model\Mail\MailTemplateVariables
-     */
-    private function createTwoFactorAuthenticationCodeMailTemplateVariables(): MailTemplateVariables
-    {
-        $mailTemplateVariables = new MailTemplateVariables(t('Two factor authentication code'));
-
-        return $mailTemplateVariables->addVariable(
-            TwoFactorAuthenticationMail::VARIABLE_AUTHENTICATION_CODE,
-            t('Authentication code'),
-            MailTemplateVariables::CONTEXT_BODY,
-            MailTemplateVariables::REQUIRED_BODY,
-        );
-    }
-
-    private function registerTwoFactorAuthenticationCodeMailTemplate(): void
-    {
-        $mailTemplateVariables = $this->createTwoFactorAuthenticationCodeMailTemplateVariables();
-
-        $this->addMailTemplateVariables(TwoFactorAuthenticationMail::TWO_FACTOR_AUTHENTICATION_CODE, $mailTemplateVariables);
     }
 }
